+++ NEWS.md
@@ -1,6 +1,8 @@
 shinyFiles 0.7.1
 --------------------------------------------------------------------
 
+* Close modal on ESC (@vnijs)
+
 * Force file selection on double click (@AFriendlyRobot [PR95](https://github.com/thomasp85/shinyFiles/pull/95))
 
 * `shinyFiles` now uses the [fs](https://github.com/r-lib/fs) package and works with Chinese, Russian, etc. characters (@vnijs [PR92](https://github.com/thomasp85/shinyFiles/pull/92))